src/styletts2/requirements.txt
SoundFile
torchaudio
munch
torch
pydub
pyyaml
librosa
nltk
matplotlib
accelerate
transformers
einops
einops-exts
tqdm
typing
typing-extensions
git+https://github.com/resemble-ai/monotonic_align.git
gruut>=2.3.4
gruut-ipa>=0.13.0
gruut-lang-en>=2.0.0